LayoutTests/http/tests/security/http-0.9/image-on-HTTP-0.9-page-blocked-expected.txt

CONSOLE MESSAGE: Sandboxing 'http://127.0.0.1:8000/security/http-0.9/resources/nph-image-on-HTTP-0.9-page-blocked.pl' because it is using HTTP/0.9.
CONSOLE MESSAGE: Cancelled resource load from 'http://127.0.0.1:8080/security/http-0.9/resources/nph-image.pl' because it is using HTTP/0.9 on a non-default port.


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
